requirements.txt
pennylane>=0.37
pennylane-lightning>=0.37
qiskit-terra>=0.46
numpy>=1.26
scipy>=1.11
scikit-image>=0.22
pillow>=10.0
torch>=2.2
torchvision>=0.17
matplotlib>=3.8
seaborn>=0.13
pandas>=2.1
pyyaml>=6.0
lpips>=0.1
wandb>=0.16
tensorboard>=2.15
click>=8.1
rich>=13.7
